Move SM utils into sm_utils.py
